✨ feat: Rota: Criação de organização

+++ apps/api/prisma/seed.ts
@@ -1,36 +1,42 @@
-import { faker } from "@/lib/faker";
 import { prisma } from "@/lib/prisma";
-import {
-  makeMembers,
-  makeOrganizations,
-  makeProjects,
-  makeUsers,
-} from "./factories";
+import { makeMembers } from "./factories/make-members";
+import { makeOrganizations } from "./factories/make-organizations";
+import { makeProjects } from "./factories/make-projects";
+import { makeUsers } from "./factories/make-users";
 
-interface Table {
+interface PgTable {
   tablename: string;
 }
 
-async function run() {
-  const tables: Table[] = await prisma.$queryRaw`
-    SELECT tablename
-    FROM pg_tables
-    WHERE schemaname = 'public'
-  `;
-
-  for (const table of tables) {
-    await prisma.$queryRawUnsafe(`
-      TRUNCATE TABLE "${table.tablename}" RESTART IDENTITY CASCADE;
-    `);
-  }
-
-  await makeUsers({ prisma, faker, length: 20 });
-  await makeOrganizations({ prisma, faker });
-  await makeMembers({ prisma, faker });
-  await makeProjects({ prisma, faker });
+interface ISeed {
+  clearTables: boolean;
+  createData: boolean;
 }
 
-run()
+async function seed({ clearTables, createData }: ISeed) {
+  if (clearTables) {
+    const tables: PgTable[] = await prisma.$queryRaw`
+      SELECT tablename
+      FROM pg_tables
+      WHERE schemaname = 'public'
+    `;
+
+    for (const table of tables) {
+      await prisma.$queryRawUnsafe(`
+        TRUNCATE TABLE "${table.tablename}" RESTART IDENTITY CASCADE;
+      `);
+    }
+  }
+
+  if (createData) {
+    await makeUsers({ length: 20 });
+    await makeOrganizations();
+    await makeMembers();
+    await makeProjects();
+  }
+}
+
+seed({ clearTables: true, createData: true })
   .then(async () => {
     await prisma.$disconnect();
   })
